requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-cors-headers==3.11.0
django-extensions==3.2.1
django-nested-admin==3.4.0
django-phonenumber-field==6.3.0
django-simple-history==3.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
phonenumbers==8.12.52
psycopg2==2.9.3
pycodestyle==2.8.0
pydot==1.4.2
pygraphviz==1.10
PyJWT==2.4.0
pyparsing==3.0.9
python-monkey-business==1.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
gunicorn==20.1.0
pip~=23.1.2
wheel~=0.38.4
setuptools~=65.5.1
requests~=2.31.0
certifi~=2023.7.22
idna~=3.4
urllib3~=2.0.4
